Modifies the way data is retrieved filtered by date when the dashboard is first created. All initially loaded graphs will be based on the availability, in date, of PACTA temperature data.
Until now, when a user enters the web and the dashboard is created for the first time, if he/she searches for a specific data and there is no data for the given date, the previous day is searched. This can cause confusion as it produces graphs with different dates from the start of the application. This does not happen during the dashboard update when a date is manually selected in the widget.